[pkg/otlp/logs] Switch to Translator design
